+++ src/OpenLoco/src/EditorController.cpp
@@ -245,23 +245,6 @@ namespace OpenLoco::EditorController
         }
     }
 
-    // 0x0043EE25
-    static bool validateStep1()
-    {
-        if (!Game::hasFlags(GameStateFlags::tileManagerLoaded))
-        {
-            return true;
-        }
-
-        if (TownManager::towns().size() >= Limits::kMinTowns)
-        {
-            return true;
-        }
-
-        GameCommands::setErrorText(StringIds::at_least_one_town_be_built);
-        return false;
-    }
-
     // 0x0043D0FA
     void goToPreviousStep()
     {
@@ -319,6 +302,26 @@ namespace OpenLoco::EditorController
         return StringIds::null;
     }
 
+    // 0x0043EE25
+    static StringId validateLandscapeEditor()
+    {
+        if (Game::hasFlags(GameStateFlags::tileManagerLoaded))
+        {
+            if (TownManager::towns().size() < Limits::kMinTowns)
+            {
+                return StringIds::at_least_one_town_be_built;
+            }
+        }
+        else
+        {
+            if (Scenario::getOptions().generator == Scenario::LandGeneratorType::PngHeightMap)
+            {
+                return StringIds::png_heightmap_must_be_generated;
+            }
+        }
+        return StringIds::null;
+    }
+
     // 0x0046F910
     static void setupMultiplayerData()
     {
@@ -334,6 +337,8 @@ namespace OpenLoco::EditorController
     // 0x0043D15D
     void goToNextStep()
     {
+        auto& options = Scenario::getOptions();
+
         switch (getCurrentStep())
         {
             case Step::null:
@@ -355,9 +360,9 @@ namespace OpenLoco::EditorController
                 Scenario::sub_4748D4();
                 Scenario::initialiseSnowLine();
                 Windows::Terraform::resetDefaultObjectIds();
-                Scenario::getOptions().editorStep = Step::landscapeEditor;
+                options.editorStep = Step::landscapeEditor;
                 Windows::LandscapeGeneration::open();
-                if ((Scenario::getOptions().scenarioFlags & Scenario::ScenarioFlags::landscapeGenerationDone) != Scenario::ScenarioFlags::none)
+                if ((options.scenarioFlags & Scenario::ScenarioFlags::landscapeGenerationDone) != Scenario::ScenarioFlags::none)
                 {
                     if (!Game::hasFlags(GameStateFlags::tileManagerLoaded))
                     {
@@ -368,30 +373,31 @@ namespace OpenLoco::EditorController
             }
             case Step::landscapeEditor:
             {
-                if (!validateStep1())
+                auto errorMessage = validateLandscapeEditor();
+                if (errorMessage != StringIds::null)
                 {
-                    Windows::Error::open(StringIds::cant_advance_to_next_editor_stage, GameCommands::getErrorText());
-                    break;
+                    Windows::Error::open(StringIds::cant_advance_to_next_editor_stage, errorMessage);
+                    return;
                 }
 
-                const auto cargoId = Scenario::getOptions().objective.deliveredCargoType;
+                const auto cargoId = options.objective.deliveredCargoType;
                 if (ObjectManager::get<CargoObject>(cargoId) == nullptr)
                 {
                     for (size_t i = 0; i < ObjectManager::getMaxObjects(ObjectType::cargo); i++)
                     {
                         if (ObjectManager::get<CargoObject>(i) != nullptr)
                         {
-                            Scenario::getOptions().objective.deliveredCargoType = static_cast<uint8_t>(i);
+                            options.objective.deliveredCargoType = static_cast<uint8_t>(i);
                             break;
                         }
                     }
                 }
 
                 WindowManager::closeAllFloatingWindows();
-                Scenario::initialiseDate(Scenario::getOptions().scenarioStartYear);
+                Scenario::initialiseDate(options.scenarioStartYear);
                 Scenario::initialiseSnowLine();
                 Windows::ScenarioOptions::open();
-                Scenario::getOptions().editorStep = Step::scenarioOptions;
+                options.editorStep = Step::scenarioOptions;
                 break;
             }
 
@@ -408,7 +414,7 @@ namespace OpenLoco::EditorController
                     break;
                 }
 
-                Scenario::getOptions().editorStep = Step::null;
+                options.editorStep = Step::null;
                 setupMultiplayerData();
 
                 auto path = fs::u8path(*res);
@@ -426,7 +432,7 @@ namespace OpenLoco::EditorController
                 if (!success)
                 {
                     Windows::Error::open(StringIds::scenario_save_failed);
-                    Scenario::getOptions().editorStep = Step::scenarioOptions;
+                    options.editorStep = Step::scenarioOptions;
                     break;
                 }